First iteration at improving simulation preparation time.
